Make availablePeopleSlots a long
Fixes #116

+++ core/src/main/java/com/rtm516/mcxboxbroadcast/core/models/session/SocialSummaryResponse.java
@@ -8,7 +8,7 @@ public record SocialSummaryResponse(
     boolean hasCallerMarkedTargetAsFavorite,
     boolean hasCallerMarkedTargetAsKnown,
     String legacyFriendStatus,
-    int availablePeopleSlots,
+    long availablePeopleSlots,
     int recentChangeCount,
     String watermark
 ) {
